+++ CHANGELOG.md
@@ -53,6 +53,28 @@
 
 ## Unreleased
 
+### Breaking changes
+
+- EventGateway CRDs (Tech Preview) which allow referencing a `Secret` through
+  `secretRef`, now require the key name to be provided.
+
+  So:
+
+  ```yaml
+  type: secretRef
+  secretRef:
+    name: my-backend-cluster-tls
+  ```
+
+  Becomes:
+
+  ```yaml
+  type: secretRef
+  secretRef:
+    name: my-backend-cluster-tls
+    key: tls.crt # Key name depends on the type of referenced Secret.
+  ```
+
 ### Added
 
 - API: expose `trafficDistribution` and `internalTrafficPolicy` in `DataPlane`,
@@ -72,6 +94,10 @@
   [#4781](https://github.com/Kong/kong-operator/pull/4781)
   - introduce `AIGatewayAgent` CRD with reconciler.
   [#4776](https://github.com/Kong/kong-operator/pull/4776)
+  - introduce `AIGatewayProvider` CRD with reconciler. Credentials can be sourced
+    from a Kubernetes `Secret`s.
+    [#4793](https://github.com/Kong/kong-operator/pull/4793)
+    [#4798](https://github.com/Kong/kong-operator/pull/4798)
 - GRPCRoute: enable GRPCRoute for on-prem gateway
   [#4364](https://github.com/Kong/kong-operator/pull/4364)